+++ src/main/scala/edu/rit/cs/dataunitylab/jsonoid/discovery/JsonoidParams.scala
@@ -6,6 +6,7 @@ import schemas.{PropertySet, PropertySets}
   *
   * @param additionalProperties whether `additionalProperties` is set to true in generated schemas
   * @param er the equivalence relation to use during the discovery process
+  * @param extendedFormats whether to include extended formats
   * @param formatThreshold the fraction of values that must match a given format for [[schemas.FormatProperty]] to consider the format valid
   * @param maxExamples the maximum number of examples to be kept for any examples property
   * @param resetFormatLength whether to reset max/min length of strings with [[schemas.FormatProperty]]
@@ -13,6 +14,7 @@ import schemas.{PropertySet, PropertySets}
 final case class JsonoidParams(
     val additionalProperties: Boolean = false,
     val er: EquivalenceRelation = EquivalenceRelations.KindEquivalenceRelation,
+    val extendedFormats: Boolean = false,
     val formatThreshold: Float = 1.0f,
     val maxExamples: Int = 100,
     val propSet: PropertySet = PropertySets.AllProperties,
@@ -24,6 +26,7 @@ final case class JsonoidParams(
     JsonoidParams(
       newAdditionalProperties,
       er,
+      extendedFormats,
       formatThreshold,
       maxExamples,
       propSet,
@@ -35,6 +38,19 @@ final case class JsonoidParams(
     JsonoidParams(
       additionalProperties,
       newER,
+      extendedFormats,
+      formatThreshold,
+      maxExamples,
+      propSet,
+      resetFormatLength
+    )
+  }
+
+  def withExtendedFormats(newExtendedFormats: Boolean): JsonoidParams = {
+    JsonoidParams(
+      additionalProperties,
+      er,
+      newExtendedFormats,
       formatThreshold,
       maxExamples,
       propSet,
@@ -46,6 +62,7 @@ final case class JsonoidParams(
     JsonoidParams(
       additionalProperties,
       er,
+      extendedFormats,
       newFormatThreshold,
       maxExamples,
       propSet,
@@ -57,6 +74,7 @@ final case class JsonoidParams(
     JsonoidParams(
       additionalProperties,
       er,
+      extendedFormats,
       formatThreshold,
       newMaxExamples,
       propSet,
@@ -68,6 +86,7 @@ final case class JsonoidParams(
     JsonoidParams(
       additionalProperties,
       er,
+      extendedFormats,
       formatThreshold,
       maxExamples,
       newPropSet,
@@ -79,6 +98,7 @@ final case class JsonoidParams(
     JsonoidParams(
       additionalProperties,
       er,
+      extendedFormats,
       formatThreshold,
       maxExamples,
       propSet,